Create package whitelist from package.xml
The prior mechanism to generate an ignore/include list was to
use a static file at the workspace root. This doesn't play very
well with our build system so instead use package.xml. Any
<build_depend> package will now be included in the ros1_bridge
build.

This is done using a simple python script to parse package.xml
which is then called by cmake (stdout of the script is the list).
We then feed that list into a modified filter_packages() function
that takes a list of include/ignore packages rather than using
the static list.