Whitelist packages (#6)
* Introduces package include and exclude lists

* Adds cmake file that parses package include resp. exclude list
* Filtering cmake dependencies based on include/exclude list

https://github.com/ros2/ros1_bridge/issues/284

Signed-off-by: norro <github@norro.de>
Co-authored-by: James Prestwood <jprestwood@locusrobotics.com>

* Passes package incl./excl. to bridge generation

* CMake passes package include/ignore lists to python scripts generating
the bridge c++ code
* Generation of factories and mappings restricted to packages filtered
by include/ignore lists

https://github.com/ros2/ros1_bridge/issues/284

Signed-off-by: norro <github@norro.de>
Co-authored-by: James Prestwood <jprestwood@locusrobotics.com>

* Create package whitelist from package.xml

The prior mechanism to generate an ignore/include list was to
use a static file at the workspace root. This doesn't play very
well with our build system so instead use package.xml. Any
<build_depend> package will now be included in the ros1_bridge
build.

This is done using a simple python script to parse package.xml
which is then called by cmake (stdout of the script is the list).
We then feed that list into a modified filter_packages() function
that takes a list of include/ignore packages rather than using
the static list.

* Replace rosidl_cmake with rosidl_common

The former is deprecated and the build was warning of this.

---------

Signed-off-by: norro <github@norro.de>
Co-authored-by: norro <github@norro.de>